Bug fix on components
There were some inconsistencies in the html structure of the
BaseLayout component since it was rendering it head element twice per
page. So I created a new component to address this issue.

I've also added the navigation component to the posts layout until I
further design something unique to it.